refactor: avoid barrel
